user_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
Flask-Cors==3.0.10
python-dotenv==1.0.0
pymysql==1.1.0
gunicorn==20.1.0
requests==2.31.0
pytest==7.4.2
Flask-JWT-Extended==4.7.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.29.0